airflow_api_etl/requirements.txt
apache-airflow>=2.7.1
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
SQLAlchemy>=1.4.50
pandas>=2.1.0
python-geohash>=0.8.0
dags>=0.1.0
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1 